Fixing race condition on tipline bot when user query submission times out
If two or more messages are sent at the same second, when the user session times out, more than one request would be created in Check. Changing to a more granular check avoids this situation.

Fixes CHECK-2022.

+++ app/models/bot/smooch.rb
@@ -950,7 +950,7 @@ class Bot::Smooch < BotUser
 
   def self.refresh_smooch_menu_timeout(message, app_id)
     uid = message['authorId']
-    time = Time.now.to_i
+    time = Time.now.to_f
     Rails.cache.write("smooch:last_message_from_user:#{uid}", time)
     self.delay_for(15.minutes).timeout_smooch_menu(time, message, app_id)
   end
